+++ bq/src/components/Footer/Footer.jsx
@@ -1,8 +1,9 @@
 import React from "react";
+import '../Footer/Footer.scss'
 
 const Footer = () => {
   return (
-    <div className="containerFooter">
+    <div className="container__footer">
       <div className="footer">
         <p>
         2021 All rights reserved - Made by Flor Jardinez  & Ana Karina Dávila.
