+++ bq/src/components/Footer/Footer.scss
@@ -1,0 +1,13 @@
+@import "../../styles/variables";
+
+.container__footer{
+    display: flex;
+    justify-content: center;
+    align-content: center;
+}
+.footer{
+    color: $text-on-primary;
+    font-family: 'Roboto', sans-serif;
+    font-weight: bold;
+    font-size: 16px;
+}
