--- bq/src/components/Login.jsx
@@ -1,42 +1,0 @@
-import React, { useState } from "react";
-import { auth, login } from "../lib/firebase";
-import { logo } from "../assets/banner.png";
-import LoginForm from "./LoginForm";
-import Footer from "./Footer";
-import { useNavigate } from 'react-router-dom';
-import "../css/App.css"
-const Login = () => {
-  const [error, setError] = useState("");
-  const navigate = useNavigate();
-  const handleSubmit = async (email, password) => {
-    const user = auth.currentUser;
-    if (user) {
-      navigate("/WaiterProfile");
-    } else {
-      try {
-        await login(email, password);
-        console.log('login exitoso')
-        navigate("/WaiterProfile");
-      } catch (error) {
-        console.error(error)
-        setError(error.message);
-        setTimeout(() => setError(""), 1500);
-      }
-    }
-  };
-  return (
-    <div className="content">
-      <div className="container__Login">
-        <header className="logo">
-          <img src={logo} alt="logoBQ" />
-        </header>
-        <div className="login__form">
-          <LoginForm handleSubmit={handleSubmit} />
-          <div className="error__section">{error}</div>
-        </div>
-      </div>
-      <Footer />
-    </div>
-  );
-};
-export default Login;
