+++ bq/src/components/Login/LoginForm.jsx
@@ -1,7 +1,6 @@
 import React, { useState } from "react";
-import { iconEmail } from "../assets/email.png";
-import { iconPassword } from "../assets/password.png";
-
+import iconEmail from "../../assets/email.png";
+import iconPassword from "../../assets/password.png";
 const LoginForm = ({ handleSubmit }) => {
   const [state, setState] = useState({});
   const handleEmail = (e) => setState({ ...state, email: e.target.value });
@@ -12,18 +11,22 @@ const LoginForm = ({ handleSubmit }) => {
 
   return (
     <form className="form">
-      <img src={ iconEmail } alt="iconEmail" />
-      <input
-        type="email"
-        placeholder="Write your Email"
-        onChange={handleEmail}
-      />
-      <img src={iconPassword} alt="iconPassword" />
-      <input
-        type="password"
-        placeholder="Write your Password"
-        onChange={handlePassword}
-      />
+      <div className="input__form">
+        <img className="icon" src={iconEmail} alt="iconEmail" />
+        <input
+          type="email"
+          placeholder="Write your Email"
+          onChange={handleEmail}
+        />
+      </div>
+      <div className="input__form">
+        <img className="icon" src={iconPassword} alt="iconPassword" />
+        <input
+          type="password"
+          placeholder="Write your Password"
+          onChange={handlePassword}
+        />
+      </div>
       <button
         className="primary-button"
         onClick={(e) => {
