+++ bq/src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "../src/routes/App";
+import "../src/styles/_main.scss";
 
 ReactDOM.render(
   <React.StrictMode>
