+++ bq/src/styles/App.scss
@@ -1,14 +1,15 @@
 @import url("https://fonts.googleapis.com/css?family=Open+Sans:300,400&display=swap");
-/* @import "../scss/variable"; */
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap');
+@import "../styles/variables"; 
 html,
 body {
   background-image: url("../assets/background.jpg");
   margin: 0;
   padding: 0;
-  font-family: "Open Sans", sans-serif;
+  font-family: 'Roboto', sans-serif;
   overflow-x: hidden !important;
   line-height: 10px;
-  color: #FEF9EF;
+  color: $text-on-primary;
 }
 .root {
   min-height: 90vh;
