+++ bq/src/styles/_main.scss
@@ -1,0 +1,4 @@
+@import "../components/Footer/Footer.scss";
+@import "../components/Login/Login.scss";
+@import "../components/WaiterProfile/WaiterProfile.scss";
+@import "../styles/App.scss";
